增加注释 优化结构等
