+++ packages/ewa/lib/ewa.js
@@ -4,18 +4,14 @@ var wxPromisify = require('./plugins/wxPromisify');
 
 var enableState = require('./plugins/enableState');
 
+var createStore = require('./plugins/reactiveStore');
+
 var mixin = require('./mixins/mixin');
-
-var watcherInstall = require('./observer/install');
-
-var _require = require('./observer/reactive'),
-    reactive = _require.reactive;
 
 var ewa = {
   mixin: mixin,
   enableState: enableState,
-  watcherInstall: watcherInstall,
-  reactive: reactive
+  createStore: createStore
 };
 wxPromisify(ewa);
 module.exports = ewa;
