+++ packages/ewa/lib/utils/removeEmptyArr.js
@@ -1,0 +1,6 @@
+"use strict";
+
+module.exports = function removeEmptyArr(obj, key) {
+  if (!obj || !Array.isArray(obj[key])) return;
+  if (obj[key].length === 0) delete obj[key];
+};
