+++ packages/ewa/package.json
@@ -27,7 +27,11 @@
     "lodash.clonedeep": "^4.5.0",
     "lodash.get": "^4.4.2",
     "lodash.keys": "^4.2.0",
-    "lodash.set": "^4.3.2"
+    "lodash.set": "^4.3.2",
+    "lodash.isobject": "^v3.0.2",
+    "lodash.isplainobject": "^v4.0.6",
+    "lodash.isfunction": "^v3.0.9",
+    "lodash.has": "^v4.5.2"
   },
   "devDependencies": {
     "@babel/cli": "^7.0.0",
