+++ packages/ewa/src/utils/isExistSameId.js
@@ -1,0 +1,7 @@
+
+module.exports = function isExistSameId(arr, id) {
+  if (Array.isArray(arr) && arr.length) {
+    return !!arr.find(item => item.id === id);
+  }
+  return false;
+}
