+++ packages/ewa/src/utils/removeById.js
@@ -1,0 +1,7 @@
+
+module.exports = function removeById(arr, id) {
+  if (Array.isArray(arr) && arr.length) {
+    return arr.filter(item => item.id !== id);
+  }
+  return arr;
+}
